requirements.txt
langchain
beautifulsoup4
faiss-cpu
langchain-google-genai
langchain-google-vertexai
langchainhub
google-cloud-bigquery